rewrite: reworked filtering process, worked on subdirectories, added documentation
Moved filtering setup into helper `gatherFilters`,
changed walker/rewriter.go so that empty subdirectories will be maintained, added
logic in rewriter to detect empty sudirectories because nothing was included,
added parameter `keepEmptyDirecory` to `filterAndReplaceSnapshot`
so rewriter knows when it deals with the `--include` variants of the filtering process.

As a consequence runRepairSnapshots had to be modified to be in line
with the include filtering process.

Added documentation for the include processing.